improvement(hdrh): rename 'csrangehistogram.py' to 'hdrhistogram.py'
We are going to add support of hdr histogram files for other
benchmarking tools too.
So, rename the 'sdcm/utils/csrangehistogram.py' module
to the 'sdcm/utils/hdrhistogram.py'.

Also, delete the 'sdcm/utils/cshdrhistogram.py' which is dead code.

(cherry picked from commit d66f2f2ce3f0923b431c5161c0cca1532cbb864d)
